+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [Unreleased]
+
+### Fixed
+
+* Fix the spell known load condition not correctly checking if the spell is known (by [novsirion])
+
 ## [1.8.1] - 2022-10-26
 
 ### Fixed
@@ -1128,3 +1134,4 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 [gitarrg]: https://github.com/gitarrg
 [tflo]: https://github.com/tflo
 [yannlugrin]: https://github.com/yannlugrin
+[novsirion]: https://github.com/novsirion