+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 ### Added
 * Update Chinese translations (by Jokey)
+* Update Evoker talent data
 
 ## [1.9.0] - 2022-11-15